+++ example/lib/presentation/recognize_speech/bloc/recognize_speech_event.dart
@@ -9,3 +9,9 @@ final class DoRecognizeEvent extends RecognizeSpeechEvent {
 
   const DoRecognizeEvent(this.text);
 }
+
+final class DoFetchFoodItemEvent extends RecognizeSpeechEvent {
+  final PassioFoodDataInfo foodDataInfo;
+
+  const DoFetchFoodItemEvent(this.foodDataInfo);
+}