+++ ios/PassioNutritionAISDK.xcframework/ios-arm64/PassioNutritionAISDK.framework/Headers/PassioNutritionAISDK-Swift.h
@@ -277,6 +277,8 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 #if __has_warning("-Watimport-in-framework-header")
 #pragma clang diagnostic ignored "-Watimport-in-framework-header"
 #endif
+@import CoreFoundation;
+@import UIKit;
 #endif
 
 #endif
@@ -297,6 +299,15 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 #endif
 
 #if defined(__OBJC__)
+
+@class NSCoder;
+
+SWIFT_CLASS("_TtC20PassioNutritionAISDK26AVCaptureViedeoPreviewView")
+@interface AVCaptureViedeoPreviewView : UIView
+- (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE;
+- (void)layoutSubviews;
+- (nonnull instancetype)initWithFrame:(CGRect)frame SWIFT_UNAVAILABLE;
+@end